app/requirements.txt
shell-genie
openai-whisper
sounddevice
soundfile
pydub
pyttsx3
